Refactoring: remove code duplicate between alerter ask/post.
