+++ examples/slack_alert/steps/evaluator.py
@@ -1,0 +1,16 @@
+import numpy as np
+from sklearn.base import ClassifierMixin
+
+from zenml.steps import step
+
+
+@step
+def evaluator(
+    X_test: np.ndarray,
+    y_test: np.ndarray,
+    model: ClassifierMixin,
+) -> float:
+    """Calculate the test set accuracy of an sklearn model."""
+    test_acc = model.score(X_test, y_test)
+    print(f"Test accuracy: {test_acc}")
+    return test_acc
