+++ examples/slack_alert/steps/importer.py
@@ -1,0 +1,16 @@
+import numpy as np
+
+from zenml.integrations.sklearn.helpers.digits import get_digits
+from zenml.steps import Output, step
+
+
+@step
+def importer() -> Output(
+    X_train=np.ndarray,
+    X_test=np.ndarray,
+    y_train=np.ndarray,
+    y_test=np.ndarray,
+):
+    """Load the digits dataset as numpy arrays."""
+    X_train, X_test, y_train, y_test = get_digits()
+    return X_train, X_test, y_train, y_test
