+++ examples/slack_alert/steps/trainer.py
@@ -1,0 +1,16 @@
+import numpy as np
+from sklearn.base import ClassifierMixin
+from sklearn.svm import SVC
+
+from zenml.steps import step
+
+
+@step
+def svc_trainer(
+    X_train: np.ndarray,
+    y_train: np.ndarray,
+) -> ClassifierMixin:
+    """Train a sklearn SVC classifier."""
+    model = SVC(gamma=0.001)
+    model.fit(X_train, y_train)
+    return model
