+++ src/zenml/alerter/__init__.py
@@ -19,10 +19,11 @@ This is useful to immediately get notified when failures happen,
 and also for general monitoring / reporting.
 """
 
-from zenml.alerter.alerter_step import alerter_step
+from zenml.alerter.alerter_step import alerter_ask_step, alerter_post_step
 from zenml.alerter.base_alerter import BaseAlerter
 
 __all__ = [
     "BaseAlerter",
-    "alerter_step",
+    "alerter_post_step",
+    "alerter_ask_step",
 ]
