+++ src/zenml/alerter/alerter_step.py
@@ -18,8 +18,29 @@ from zenml.steps import StepContext, step
 from zenml.steps.step_interfaces.base_alerter_step import BaseAlerterStepConfig
 
 
+def _check_alerter_registered(context: StepContext):
+    """
+    Raise an error if the active stack has no alerter registered.
+    """
+    # TODO: duplicate code with examples/feast_feature_store/run.py
+    if not context.stack:
+        raise DoesNotExistException(
+            "No active stack is available. "
+            "Please make sure that you have registered and set a stack."
+        )
+    if not context.stack.alerter:
+        raise DoesNotExistException(
+            "The active stack needs to have an alerter component registered "
+            "to be able to use `alerter_step`. "
+            "You can create a new stack with e.g. a Slack alerter component or update "
+            "your existing stack to add this component, e.g.:\n\n"
+            "  'zenml alerter register slack_alerter --flavor=slack' ...\n"
+            "  'zenml stack register stack-name -al slack_alerter ...'\n"
+        )
+
+
 @step
-def alerter_step(
+def alerter_post_step(
     config: BaseAlerterStepConfig, context: StepContext, message: str
 ) -> bool:
     """Post a given message to the registered alerter component of the
@@ -34,23 +55,31 @@ def alerter_step(
         True if operation succeeded, else False.
 
     Raises:
-        ValueError if active stack has no slack alerter.
+        DoesNotExistException if active stack has no slack alerter.
     """
-
-    # TODO: duplicate code with examples/feast_feature_store/run.py
-    if not context.stack:
-        raise DoesNotExistException(
-            "No active stack is available. "
-            "Please make sure that you have registered and set a stack."
-        )
-    if not context.stack.alerter:
-        raise ValueError(
-            "The active stack needs to have an alerter component registered "
-            "to be able to use `alerter_step`. "
-            "You can create a new stack with e.g. a Slack alerter component or update "
-            "your existing stack to add this component, e.g.:\n\n"
-            "  'zenml alerter register slack_alerter --flavor=slack' ...\n"
-            "  'zenml stack register stack-name -al slack_alerter ...'\n"
-        )
-
+    _check_alerter_registered(context)
     return context.stack.alerter.post(message, config)
+
+
+@step
+def alerter_ask_step(
+    config: BaseAlerterStepConfig, context: StepContext, message: str
+) -> bool:
+    """Post a given message to the registered alerter component of the
+    active stack and wait for approval.
+    This can be useful, e.g. to easily get a human in the loop when
+    deploying models.
+
+    Args:
+        config: Runtime configuration for the slack alerter.
+        context: StepContext of the ZenML repository.
+        message: Initial message to be posted.
+
+    Returns:
+        True if a user approved the operation, else False
+
+    Raises:
+        ValueDoesNotExistExceptionError if active stack has no slack alerter.
+    """
+    _check_alerter_registered(context)
+    return context.stack.alerter.ask(message, config)
